fix: 메인페이지 접속시 상태 초기화 안되는 이슈 - 메인 페이지에 접속시 상태를 초기화 할수 있도록 구현함.
